app/requirements.txt
confluent_kafka
sqlalchemy
PyMySQL
python-dotenv
ujson
kafka-python
requests
flask
aiohttp
redis